[LCR-14] refactor: PrincipalDetails 패키지 경로 변경
